+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Support output block in tag line.
+
 ## 0.2.0 - 2022-01-25
 
 ### Added